+++ packages/lint-staged-config/.lintstagedrc.js
@@ -1,4 +1,4 @@
 module.exports = {
   "!(*lintstagedrc).{js,jsx,ts,tsx}": ["prettier --write", "eslint --fix"],
-  "*.{json,md,css,scss}": ["prettier --write"],
+  "*.{json,md,mdx,css,scss}": ["prettier --write"],
 }