+++ packages/wethegit-components/src/components/back-to-top/back-to-top.stories.tsx
@@ -1,6 +1,5 @@
-import { expect } from "@storybook/jest"
+import { expect, userEvent, within, waitFor } from "@storybook/test"
 import type { Meta, StoryObj } from "@storybook/react"
-import { userEvent, within, waitFor } from "@storybook/testing-library"
 
 import { BackToTop, Text } from "@local/components"
 
@@ -13,14 +12,6 @@ const meta: Meta<typeof BackToTop> = {
     onComplete: undefined,
     pixelsPerSecond: 2000,
     revealThreshold: "85vh",
-  },
-  argTypes: {
-    easingFunction: {
-      control: "function",
-    },
-    onComplete: {
-      control: "function",
-    },
   },
 }