+++ packages/wethegit-components/src/components/back-to-top/back-to-top.tsx
@@ -1,3 +1,5 @@
+"use client"
+
 import { useCallback, useRef } from "react"
 import { useUserPrefs, useInView } from "@wethegit/react-hooks"
 
@@ -42,7 +44,7 @@ export function BackToTop({
 }: BackToTopProps): JSX.Element {
   const buttonRef = useRef<HTMLButtonElement>(null)
   const { prefersReducedMotion } = useUserPrefs()
-  const [setReferenceRef, referenceIsInView] = useInView(1, false, false)
+  const [setReferenceRef, referenceIsInView] = useInView(0, false, false)
 
   const handleClick = useCallback(
     (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => {
@@ -54,10 +56,10 @@ export function BackToTop({
 
       const position = window.scrollY
       const duration = prefersReducedMotion ? 0 : getDuration(position, pixelsPerSecond)
-      let starttime: null | number = null
+      let starttime: number | null = null
 
       const animate = (delta: number) => {
-        if (starttime === null) starttime = 0
+        if (starttime === null) starttime = delta
 
         const playHead = clamp(0, 1, (delta - starttime) / duration)
         const easedPlayHead = easingFunction
@@ -76,7 +78,14 @@ export function BackToTop({
 
       requestAnimationFrame(animate)
     },
-    [prefersReducedMotion]
+    [
+      easingFunction,
+      focusOnCompleteCssSelector,
+      onComplete,
+      pixelsPerSecond,
+      prefersReducedMotion,
+      props.onClick,
+    ]
   )
 
   return (
@@ -92,7 +101,7 @@ export function BackToTop({
         ref={buttonRef}
         className={classnames([
           styles.button,
-          (!referenceIsInView || revealThreshold.startsWith("0")) && styles.buttonShown,
+          (revealThreshold.startsWith("0") || !referenceIsInView) && styles.buttonShown,
           className,
         ])}
         onClick={handleClick}