+++ packages/wethegit-components/src/components/in-view/in-view.tsx
@@ -1,10 +1,12 @@
-import { createContext, ElementType, useMemo } from "react"
+"use client"
+
+import { createContext, ElementType } from "react"
 import { useInView } from "@wethegit/react-hooks"
 
 import { Tag } from "@local/components"
 import type { TagProps } from "@local/components"
 
-export type InViewContext = {
+export type InViewContextReturn = {
   isInView: boolean
   domNode: HTMLElement | undefined
 }
@@ -12,52 +14,36 @@ export type InViewContext = {
 export type InViewProps<TAs extends ElementType> = TagProps<TAs> & {
   /**
    * Percentage of the entry within the viewport for it to be considered in
-   * view; a number between 0 and 1.
+   * view; a number between 0 and 1. Also accepts an IntersectionObserver options
+   * object.
    */
-  threshold?: number
+  observerOptions?: number | IntersectionObserverInit
   /**
-   * Unobserve the entry once it is considered to be in view.
+   * Whether to detach the observer from the DOM element after the first
+   * intersection callback is invoked.
    */
   once?: boolean
   /**
-   * Consider the entry "in view" if it is above the viewport when the page loads
-   * initially. This is most often used for ensuring that content is already
-   * animated in, when loading a document from a scroll position other than the top.
+   * Whether to consider the element already "in-view", if it is already scrolled
+   * beyond the bounds of the viewport when the element is mounted.
    */
   setInViewIfScrolledPast?: boolean
-  /**
-   * Ensure that the entry remains in view until it fully exits the viewport,
-   * regardless of the threshold. Note that this will not have any effect within
-   * an `<iframe>`.
-   */
-  matchRootMarginToThreshold?: boolean
 }
 
-export const InViewContext = createContext<InViewContext>({
+export const InViewContext = createContext<InViewContextReturn>({
   isInView: false,
   domNode: undefined,
 })
 
-export function InView<TAs extends ElementType>({
-  threshold = 0.3,
+export function InView<TAs extends ElementType, T extends HTMLElement>({
+  observerOptions = 0.3,
   once = false,
   setInViewIfScrolledPast = false,
-  matchRootMarginToThreshold = true,
   ...props
 }: InViewProps<TAs>) {
   const { as = "div", className, ...rest } = props
 
-  const observerOptions = useMemo(
-    () => ({
-      threshold,
-      ...(matchRootMarginToThreshold && {
-        rootMargin: `${threshold * 100}% 0px 0px 0px`,
-      }),
-    }),
-    [matchRootMarginToThreshold, threshold]
-  )
-
-  const [setRef, isInView, domNode] = useInView(
+  const [setRef, isInView, domNode] = useInView<T>(
     observerOptions,
     once,
     // only honor "setInViewIfScrolledPast" if "once" is true: