+++ packages/wethegit-components/src/components/modal/modal.stories.tsx
@@ -1,6 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react"
-import { within, userEvent, waitFor } from "@storybook/testing-library"
-import { expect } from "@storybook/jest"
+import { within, userEvent, waitFor, expect } from "@storybook/test"
 import { UserPreferencesProvider } from "@wethegit/react-hooks"
 import { useModal } from "@wethegit/react-modal"
 import { useRef } from "react"
@@ -30,13 +29,19 @@ type Story = StoryObj<typeof Modal>
 
 export const Default: Story = {
   render: (args) => {
+    const modalRootRef = useRef<HTMLDivElement>(null)
     return (
-      <Modal {...args} appendToBody={false}>
-        <p style={{ color: "black", margin: 0 }}>
-          <strong>Close</strong> button doesn't work becayse inside the story we use the{" "}
-          <code>isOpen</code> control.
-        </p>
-      </Modal>
+      <>
+        <div ref={modalRootRef}></div>
+        {modalRootRef.current && (
+          <Modal {...args} renderTo={modalRootRef.current}>
+            <p style={{ color: "black", margin: 0 }}>
+              <strong>Close</strong> button doesn't work because inside the story we use
+              the <code>isOpen</code> control.
+            </p>
+          </Modal>
+        )}
+      </>
     )
   },
 }
@@ -44,6 +49,7 @@ export const Default: Story = {
 export const WithButtonTrigger: Story = {
   render: () => {
     const triggerRef = useRef<HTMLButtonElement>(null)
+    const modalRootRef = useRef<HTMLDivElement>(null)
     const { isOpen, toggle } = useModal({
       triggerRef,
     })
@@ -70,9 +76,17 @@ export const WithButtonTrigger: Story = {
         >
           {isOpen ? "Close" : "Open"} Modal
         </button>
-        <Modal isOpen={isOpen} toggle={toggle} aria-label="My Modal">
-          <p style={{ color: "black", margin: 0 }}>Hey! The modal is open!</p>
-        </Modal>
+        <div ref={modalRootRef}></div>
+        {modalRootRef.current && (
+          <Modal
+            renderTo={modalRootRef.current}
+            isOpen={isOpen}
+            toggle={toggle}
+            aria-label="My Modal"
+          >
+            <p style={{ color: "black", margin: 0 }}>Hey! The modal is open!</p>
+          </Modal>
+        )}
       </div>
     )
   },